refactor cameras
