+++ airo-camera-toolkit/docs/live_charuco_pose.py
@@ -1,6 +1,6 @@
 import cv2
 from airo_camera_toolkit.calibration.fiducial_markers import detect_charuco_board, draw_frame_on_image
-from airo_camera_toolkit.cameras.zed2i import Zed2i
+from airo_camera_toolkit.cameras.zed.zed2i import Zed2i
 from airo_camera_toolkit.utils.image_converter import ImageConverter
 
 camera = Zed2i(fps=30)
